requirements.txt
google-generativeai
Pillow
python-dotenv
opencv-python
gradio_client
polars
inference-gpu[yolo-world]==0.9.12rc1
supervision==0.19.0rc3